backend/requirements.txt
# AI-HCP CRM Backend Requirements
# Healthcare Professional Interaction Logging Module

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.12.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment Variables
python-dotenv>=1.0.0

# LangGraph - AI Workflow
langgraph>=0.0.30

# Groq - LLM API
groq>=0.4.0

# Async Support
httpx>=0.25.0

# JSON Support
orjson>=3.9.0

# DateTime handling
python-dateutil>=2.8.0